Crate ab_proof_of_space

Proof of space implementation

Modules

chia

Chia proof of space implementation

chiapos

Chia proof of space reimplementation in Rust

shim

Shim proof of space implementation that works much faster than Chia and can be used for testing purposes to reduce memory and CPU usage

Structs

PosProofs

Proof-of-space proofs

Enums

PosTableType

Proof of space table type

Traits

Table

Proof of space kind

TableGenerator

Stateful table generator with better performance.