ab_proof_of_space

Module chiapos

Chia proof of space reimplementation in Rust

Structs

Proofs

Proof-of-space proofs

Tables

Collection of Chia tables

Constants

PROOF_SIZE

Size of the proof in bytes for a given K value

Traits

SupportedKValue

Supported K values for Chia proof of space tables